+++ model_navigator/runners/torch.py
@@ -14,7 +14,6 @@
 """Torch runners."""
 
 import gc
-from copy import deepcopy
 from typing import List, Optional
 
 from model_navigator.configuration import Format, TensorType
@@ -341,16 +340,11 @@ class TorchCompileCUDARunner(_BaseTorchRunner):
     def activate_impl(self):
         """Runner activation implementation."""
         super().activate_impl()
-        model_copy = deepcopy(self._loaded_model)
-        if not self._inplace:
-            # offload original model from the gpu so other processes can use the memory
-            self.model.to("cpu")
-        model_copy.to(self.device).eval()
         LOGGER.info(
             f"Using torch.compile with config: fullgraph={self.fullgraph}, dynamic={self.dynamic}, backend={self.backend}, mode={self.mode}, options={self.options}"
         )
         self._loaded_model = torch.compile(
-            model=model_copy,
+            model=self._loaded_model,
             fullgraph=self.fullgraph,
             dynamic=self.dynamic,
             backend=self.backend,
@@ -361,10 +355,12 @@ class TorchCompileCUDARunner(_BaseTorchRunner):
     def deactivate_impl(self):
         """Deactivation implementation."""
         super().deactivate_impl()
-        torch._dynamo.reset()
         # offload the model from the gpu so other processes can use the memory
         if not self._inplace:
             self.model.to(self._input_module_device)
+
+        if is_torch2_available():
+            torch._dynamo.reset()
         torch.cuda.empty_cache()
         gc.collect()
 
@@ -387,8 +383,7 @@ class TorchCompileCPURunner(_BaseTorchRunner):
     def activate_impl(self):
         """Runner activation implementation."""
         super().activate_impl()
-        model_copy = deepcopy(self._loaded_model)
-        self._loaded_model = torch.compile(model_copy)
+        self._loaded_model = torch.compile(self._loaded_model)
 
     def deactivate_impl(self):
         """Deactivation implementation."""